third_party/WebKit/LayoutTests/fast/dom/SelectorAPI/id-fastpath-strict-expected.txt

This tests that the querySelector and querySelectorAll fast path for IDs is not overzelous.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS document.querySelector('#testId') is correctNode
PASS document.querySelector('div#testId') is correctNode
PASS document.querySelector('ul#testId') is null
PASS document.querySelector('ul #testId') is null
PASS document.querySelector('#testId[attr]') is null
PASS document.querySelector('#testId:not(div)') is null
PASS document.querySelectorAll('div#testId').length is 1
PASS document.querySelectorAll('div#testId').item(0) is correctNode
PASS document.querySelectorAll('#testId').length is 1
PASS document.querySelectorAll('#testId').item(0) is correctNode
PASS document.querySelectorAll('ul#testId').length is 0
PASS document.querySelectorAll('ul #testId').length is 0
PASS document.querySelectorAll('#testId[attr]').length is 0
PASS document.querySelectorAll('#testId:not(div)').length is 0
PASS successfullyParsed is true

TEST COMPLETE
